requirements.txt
bootstrap4==0.1.0
certifi==2019.6.16
cloudinary==1.17.0
dj-database-url==0.5.0
Django==2.2.3
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
Faker==2.0.0
gunicorn==19.9.0
mock==3.0.5
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
whitenoise==4.1.3
